+++ Example/ExampleTests/Fixtures/NiblessHeaderFooterView.swift
@@ -1,0 +1,19 @@
+//
+//  NiblessHeaderFooterView.swift
+//  DTCollectionViewManager
+//
+//  Created by Denys Telezhkin on 31.10.15.
+//  Copyright © 2015 Denys Telezhkin. All rights reserved.
+//
+
+import Foundation
+import UIKit
+import DTCollectionViewManager
+import DTModelStorage
+
+class NiblessHeaderFooterView : UICollectionReusableView, ModelTransfer
+{
+    func updateWithModel(model: Int) {
+        
+    }
+}