+++ Example/ExampleTests/Helpers/DTTestCollectionController.swift
@@ -19,14 +19,3 @@ class DTSupplementaryTestCollectionController: UIViewController, DTCollectionVie
     
     var collectionView : UICollectionView? = UICollectionView(frame: CGRect(x: 0, y: 0, width: 1024, height: 768), collectionViewLayout: UICollectionViewFlowLayout())
 }
-
-extension DTSupplementaryTestCollectionController : UICollectionViewDelegateFlowLayout
-{
-//    func collectionView(collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, referenceSizeForHeaderInSection section: Int) -> CGSize {
-//        return CGSize(width: 200, height: 300)
-//    }
-//    
-//    func collectionView(collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, referenceSizeForFooterInSection section: Int) -> CGSize {
-//        return CGSize(width: 200, height: 300)
-//    }
-}